refactor(test): migrate agent delegation to a continuous suite
Continues the tsx migration (CLAUDE.md: suites run via tsx, there is no
Vitest runner). Straight port, all 15 tests preserved.

  agentDelegation -> test:agent-delegation (15)

test:agent-runtime:vitest batched four files; agentDelegation is migrated
here so the script now names the three that remain, and the new suite is
wired into test:unit alongside it.

One structural change worth noting: the original reached into
`createWorkerInstance.mock.results[0].value` to inspect the delegated
worker. The factory now records the workers it creates, so the assertion
reads a plain array instead of vitest mock internals.
